Updating csv modeling and readme

+++ etna/lib/etna/spec/vcr.rb
@@ -71,7 +71,9 @@ def setup_base_vcr(spec_helper_dir, server: nil, application: nil)
       end
     end
 
-    c.debug_logger = File.open('log/vcr_debug.log', 'w')
+    if File.exists?('log')
+      c.debug_logger = File.open('log/vcr_debug.log', 'w')
+    end
 
     c.default_cassette_options = {
         serialize_with: :compressed,
